# api/openapi.yaml
openapi: 3.1.0
info:
  title: local-generative-ai gateway
  version: 0.2.2
  summary: OpenAI-compatible API gateway in front of a self-hosted open-weights coding model.
  description: |
    **Public listener** (`LGAI_ADDR`, default `127.0.0.1:8080`; containers set `:8080`):
    * `POST /v1/chat/completions`, `GET /v1/models`: OpenAI-compatible. They work with the
      official `openai` client by changing `base_url` and `api_key`.
    * `POST /v1/tasks/{task}`: coding tasks (`explain`, `review`, `tests`, `fix`). They return
      the same `chat.completion` / `chat.completion.chunk` shapes (ADR 0003).
    * `GET /healthz`, `GET /readyz`: unauthenticated probes.

    **Admin listener** (`LGAI_ADMIN_ADDR`, default `127.0.0.1:8081`): every `/admin/*` path.
    It only starts when `LGAI_ADMIN_TOKEN` is set. The public listener has **no** `/admin`
    routes: every `/admin/*` path there returns 404, unless `LGAI_ADMIN_ON_PUBLIC=true`
    explicitly mounts them (the gateway logs a WARN at startup).

    **Conventions**
    * Keys are accepted only as `Authorization: Bearer lgai_…`; never in the query string,
      a cookie or `X-API-Key`. A missing, malformed, unknown, revoked or expired key gets the
      same 401.
    * All errors use the OpenAI envelope `{"error":{"message","type","code","param"}}`,
      including 404 for unknown routes and 405 for wrong methods (the only exceptions are
...
    * Responses with 429 and 503 carry `Retry-After` (integer seconds). The `openai` Python
...
servers:
...
security:
  - apiKey: []
...
paths:
...
  /admin/keys:
    servers: &adminServers
      - url: http://127.0.0.1:8081
        description: Admin listener (reach it over SSH, e.g. ssh -L 8081:127.0.0.1:8081)
    post:
      tags: [admin]
      operationId: createKey
      summary: Create an API key. The plaintext key is returned once, in this response only.
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/KeyCreateRequest'}
            examples:
              seat:
                value: {name: lab-07, expires_at: '2026-10-14T18:00:00Z'}
              loadtest:
                summary: Load-test key (no rate limit or quota, 8 concurrent). Revoke right after the run.
                value: {name: loadtest, rpm_limit: 0, daily_token_quota: 0, max_inflight: 8}
      responses:
        '201':
          description: |
            Created. `key` is shown only here. It is stored as SHA-256 and cannot be retrieved again.
          headers:
            X-Request-ID: {$ref: '#/components/headers/X-Request-ID'}
            Cache-Control:
              schema: {type: string, const: no-store}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/KeyCreated'}
              example:
                id: 7
                name: lab-07
                key: lgai_Xy3kQ9aaAp8_sdNu-uQiIrf97dOOYJVvCVvgWjZqvec
                key_prefix: lgai_Xy3kQ9a
                created_at: '2026-10-14T13:05:00Z'
                expires_at: '2026-10-14T18:00:00Z'
                revoked_at: null
                rpm_limit: null
                daily_token_quota: null
                max_inflight: null
                last_used_at: null
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/AdminUnauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
    get:
      tags: [admin]
      operationId: listKeys
      summary: List all keys (active, expired and revoked). Never includes plaintext or hashes.
      security:
        - adminToken: []
      responses:
        '200':
          description: Keys ordered by id
          headers:
            X-Request-ID: {$ref: '#/components/headers/X-Request-ID'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/KeyList'}
              example:
                object: list
                data:
                  - {id: 1, name: team-demo, key_prefix: lgai_Q1w2E3r, created_at: '2026-10-13T09:00:00Z', expires_at: null, revoked_at: null, rpm_limit: 120, daily_token_quota: 0, max_inflight: 4, last_used_at: '2026-10-14T13:10:12Z'}
                  - {id: 7, name: lab-07, key_prefix: lgai_Xy3kQ9a, created_at: '2026-10-14T13:05:00Z', expires_at: '2026-10-14T18:00:00Z', revoked_at: '2026-10-14T17:00:00Z', rpm_limit: null, daily_token_quota: null, max_inflight: null, last_used_at: '2026-10-14T16:58:40Z'}
        '401': {$ref: '#/components/responses/AdminUnauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /admin/keys/batch:
    servers: *adminServers
    post:
      tags: [admin]
      operationId: createKeyBatch
      summary: Create `count` keys named name_prefix + zero-padded index; plaintext keys returned once.
      description: |
        Names are `name_prefix` followed by the index, zero-padded to at least 2 digits (or to the
        width of `count`): `lab-01` … `lab-40`. Use seat labels, never real names. `scripts/mint-keys.sh`
        wraps this call and writes `keys.csv` (mode 0600) plus printable slips.
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/KeyBatchRequest'}
            example: {count: 40, name_prefix: lab-, expires_at: '2026-10-14T18:00:00Z'}
      responses:
        '201':
          description: 'The created keys, each with its plaintext `key`. Sent with `Cache-Control: no-store`.'
          headers:
            X-Request-ID: {$ref: '#/components/headers/X-Request-ID'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/KeyBatchCreated'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/AdminUnauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /admin/keys/revoke:
    servers: *adminServers
    post:
      tags: [admin]
      operationId: revokeKeys
      summary: Revoke every active key whose name starts with name_prefix, or all keys.
      description: 'Send exactly one of `{"name_prefix": "lab-"}` or `{"all": true}`. Takes effect on the next request.'
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RevokeRequest'}
            examples:
              prefix: {value: {name_prefix: lab-}}
              all: {value: {all: true}}
      responses:
        '200':
          description: Number of keys revoked by this call
          headers:
            X-Request-ID: {$ref: '#/components/headers/X-Request-ID'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RevokeResult'}
              example: {revoked: 40}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/AdminUnauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /admin/keys/{id}:
    servers: *adminServers
    get:
      tags: [admin]
      operationId: getKey
      summary: One key (never plaintext or hash).
      security:
        - adminToken: []
      parameters:
        - {$ref: '#/components/parameters/KeyID'}
      responses:
        '200':
          description: The key
          headers:
            X-Request-ID: {$ref: '#/components/headers/X-Request-ID'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Key'}
        '401': {$ref: '#/components/responses/AdminUnauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/KeyNotFound'}
    delete:
      tags: [admin]
      operationId: revokeKey
      summary: Revoke a key (soft delete, idempotent). It takes effect on the next request.
      description: |
        Sets `revoked_at`. Usage rows are kept. Requests already in flight finish. Every later request
        with this key returns 401.
      security:
        - adminToken: []
      parameters:
        - {$ref: '#/components/parameters/KeyID'}
      responses:
        '200':
          description: The key after revocation
          headers:
            X-Request-ID: {$ref: '#/components/headers/X-Request-ID'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Key'}
        '401': {$ref: '#/components/responses/AdminUnauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/KeyNotFound'}
...
  /admin/maintenance:
    servers: *adminServers
    get:
      tags: [admin]
      operationId: getMaintenance
      summary: Whether maintenance mode is on.
      security:
        - adminToken: []
      responses:
        '200':
          description: Current state
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Maintenance'}
        '401': {$ref: '#/components/responses/AdminUnauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    put:
      tags: [admin]
      operationId: setMaintenance
      summary: Switch maintenance mode (the kill switch).
      description: |
        While enabled, chat and task requests return 503 `maintenance` with `Retry-After`;
        `/v1/models`, the health probes and the admin API keep working. The state is in memory
        and resets to off on restart.
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Maintenance'}
            example: {enabled: true}
      responses:
        '200':
          description: New state
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Maintenance'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/AdminUnauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: 'lgai_<43 chars base64url>'
      description: |
        Per-person API key, created by the admin. Send it only as `Authorization: Bearer lgai_...`
        (scheme case-insensitive). Query strings, cookies and `X-API-Key` are ignored.
    adminToken:
      type: http
      scheme: bearer
      description: |
        Value of `LGAI_ADMIN_TOKEN`: at least 43 characters and 16 distinct characters, no
        placeholder words (generate with `openssl rand -base64 32` or `gateway gen-secret`).
        Missing header → 401; wrong token (including a user key) → 403. Without a token the admin
        listener does not start.
  parameters:
    KeyID:
      name: id
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
  headers:
    X-Request-ID:
      description: |
        Request id, also in logs, usage rows and the upstream request. A valid incoming `X-Request-ID`
        (`[A-Za-z0-9._-]{1,64}`) is reused; otherwise 16 hex characters are generated.
      schema: {type: string}
...
  responses:
    BadRequest:
      description: Malformed JSON, missing or invalid field, unknown field (tasks, admin), unsupported parameter, or the upstream rejected the request (e.g. context length).
      headers:
        X-Request-ID: {$ref: '#/components/headers/X-Request-ID'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          examples:
            invalidJSON:
              value: {error: {message: invalid JSON body, type: invalid_request_error, code: invalid_json, param: null}}
            missingField:
              value: {error: {message: 'language is required: one of python, java, go, c, cpp', type: invalid_request_error, code: missing_field, param: language}}
            unsupported:
              value: {error: {message: 'n must be 1; multiple choices are not supported', type: invalid_request_error, code: unsupported_parameter, param: n}}
            mediaPart:
              value: {error: {message: only text content parts are supported, type: invalid_request_error, code: unsupported_parameter, param: 'messages[0].content[1].type'}}
            contextLength:
              value: {error: {message: "upstream rejected the request: maximum context length exceeded", type: invalid_request_error, code: context_length_exceeded, param: null}}
...
    AdminUnauthorized:
      description: Missing admin bearer token.
      headers:
        X-Request-ID: {$ref: '#/components/headers/X-Request-ID'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {error: {message: missing admin token, type: authentication_error, code: missing_admin_token, param: null}}
    Forbidden:
      description: Admin bearer token present but wrong (including a normal `lgai_` key).
      headers:
        X-Request-ID: {$ref: '#/components/headers/X-Request-ID'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {error: {message: invalid admin token, type: permission_error, code: forbidden, param: null}}
...
    KeyNotFound:
      description: No key with this id.
      headers:
        X-Request-ID: {$ref: '#/components/headers/X-Request-ID'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {error: {message: key not found, type: not_found_error, code: key_not_found, param: id}}
...
    PayloadTooLarge:
      description: Body larger than `LGAI_MAX_BODY_BYTES`, chat input larger than `LGAI_CHAT_MAX_INPUT_BYTES`, or task input (code + error + instructions + framework) larger than `LGAI_TASK_MAX_INPUT_BYTES`.
      headers:
        X-Request-ID: {$ref: '#/components/headers/X-Request-ID'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {error: {message: 'task input is 40213 bytes; the limit is 16384', type: invalid_request_error, code: request_too_large, param: code}}
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type, code, param]
          properties:
            message: {type: string, description: Human-readable. Never contains prompt content, keys or internal addresses.}
            type:
              type: string
              enum: [invalid_request_error, authentication_error, permission_error, not_found_error, rate_limit_error, insufficient_quota, server_error]
            code:
              type: string
              description: Machine-readable code.
              enum:
                - invalid_json
                - missing_field
                - invalid_value
                - unknown_field
                - unsupported_parameter
                - context_length_exceeded
                - request_too_large
                - request_timeout
                - invalid_api_key
                - missing_admin_token
                - forbidden
                - not_found
                - method_not_allowed
                - unknown_task
                - model_not_found
                - key_not_found
                - rate_limit_exceeded
                - concurrency_limit_exceeded
                - insufficient_quota
                - upstream_error
                - upstream_unavailable
                - upstream_timeout
                - server_overloaded
                - maintenance
                - internal_error
            param: {type: ['string', 'null'], description: Offending field, if any.}
...
    KeyCreateRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: {type: string, minLength: 1, maxLength: 100, description: 'A seat or role label such as `lab-07` or `team-demo`, never a real name. Does not have to be unique.'}
        expires_at: {type: ['string', 'null'], format: date-time, description: 'Must be in the future. Absent or null → now + `LGAI_KEY_TTL` (if > 0), else no expiry.'}
        rpm_limit: {type: ['integer', 'null'], minimum: 0, description: 'Override `LGAI_RATE_RPM` (0 = unlimited, null = default).'}
        daily_token_quota: {type: ['integer', 'null'], minimum: 0, description: 'Override `LGAI_TOKEN_QUOTA_DAILY` (0 = unlimited, null = default).'}
        max_inflight: {type: ['integer', 'null'], minimum: 0, description: 'Override `LGAI_KEY_MAX_INFLIGHT` (0 = unlimited, null = default).'}
    KeyBatchRequest:
      type: object
      additionalProperties: false
      required: [count, name_prefix]
      properties:
        count: {type: integer, minimum: 1, maximum: 200}
        name_prefix: {type: string, minLength: 1, maxLength: 90, description: e.g. `lab-`}
        expires_at: {type: ['string', 'null'], format: date-time, description: As in KeyCreateRequest.}
    Key:
      type: object
      required: [id, name, key_prefix, created_at, expires_at, revoked_at, rpm_limit, daily_token_quota, max_inflight, last_used_at]
      properties:
        id: {type: integer, format: int64}
        name: {type: string}
        key_prefix: {type: string, description: First 12 characters of the key, for identification.}
        created_at: {type: string, format: date-time}
        expires_at: {type: ['string', 'null'], format: date-time, description: An expired key gets the same 401 as an unknown one.}
        revoked_at: {type: ['string', 'null'], format: date-time}
        rpm_limit: {type: ['integer', 'null']}
        daily_token_quota: {type: ['integer', 'null']}
        max_inflight: {type: ['integer', 'null']}
        last_used_at: {type: ['string', 'null'], format: date-time}
    KeyCreated:
      allOf:
        - $ref: '#/components/schemas/Key'
        - type: object
          required: [key]
          properties:
            key:
              type: string
              pattern: '^lgai_[A-Za-z0-9_-]{43}$'
              description: Plaintext key. Shown once. Give it to the person, then forget it.
    KeyBatchCreated:
      type: object
      required: [object, data]
      properties:
        object: {type: string, const: list}
        data:
          type: array
          items: {$ref: '#/components/schemas/KeyCreated'}
    KeyList:
      type: object
      required: [object, data]
      properties:
        object: {type: string, const: list}
        data:
          type: array
          items: {$ref: '#/components/schemas/Key'}
    RevokeRequest:
      type: object
      additionalProperties: false
      description: Exactly one of `name_prefix` (non-empty) or `all` (true).
      properties:
        name_prefix: {type: string, minLength: 1}
        all: {type: boolean}
    RevokeResult:
      type: object
      required: [revoked]
      properties:
        revoked: {type: integer, description: Keys revoked by this call (already revoked ones are not counted).}
    Maintenance:
      type: object
      required: [enabled]
      additionalProperties: false
      properties:
        enabled: {type: boolean}
...